Update dependencies and plugin registrations in Flutter project
This commit updates the dependencies in `pubspec.yaml` to their latest versions, including `intl`, `http`, `flutterflow_ui`, and others. It also updates the plugin registrations in `GeneratedPluginRegistrant.swift` for macOS and `generated_plugin_registrant.cc` for Windows to include the `UrlLauncherPlugin` for URL launching functionality.

The `pubspec.lock` file has also been updated to reflect the new versions of the dependencies.